Create native zuulv3 job
So, a few things here. Currently, we'll consume 5 nodes from nodepool.
This isn't the best option, but right now we have no other way to
setup ansible_host in the inventory file.  What we really need, is the
ability to request a single node from nodepool, and then reuses the
ip we get for series of inventory hosts we list out.

Change-Id: Ieddfb0d45420749d6e9bc9451e6093284893bca2
Signed-off-by: Paul Belanger <pabelanger@redhat.com>