Replace native zuul jobs with tox based jobs
It's becoming more difficult to support a deployment installer via
zuul-executor. There is some functionality that we want to use in
ansible, but doesn't yet exist from zuul-executors. This includes
groups of groups in inventory files, group_vars next to inventory
files, and newer versions of ansible.

While I'm confident zuul will be able to do this in the future, we can
switch to tox based jobs for now with the goal to migrate back.

Change-Id: Iaace91a631dfc9528a07024c680d51e6d5a3d52f
Signed-off-by: Paul Belanger <pabelanger@redhat.com>