Use separate venv for bindep
Because we want to use bindep to manage our build dependencies for
ansible, we need to first create a virtualenv with only bindep, so we
can add the missing packages needed to build wheels for ansible.

This also adds ssl development headers, which were missing.

Remove bindep from requirements.txt, tox.ini includes it for bindep
environment.

Change-Id: I0e15a1fd97b0585c37a11044c7ad22985526cadb
Signed-off-by: Paul Belanger <pabelanger@redhat.com>