Add the ability to disable hosts in inventory
It is possible, you don't want to run windmill on a specific host.
Rather then removing the host from the inventory file, you can now place
it in the disabled inventory group and our playbooks will skip it.

This logic is used by the opendev team when running ansible-playbooks
against their control plane.

Change-Id: I717fc4b951c16c5df6286bfe13f0fe5701cf7abb
Signed-off-by: Paul Belanger <pabelanger@redhat.com>